requirements.txt
dj-database-url==0.4.1
Django==1.10.3
django-braces==1.10.0
django-environ==0.4.1
django-model-utils==2.6
djangorestframework==3.5.3
gunicorn==19.6.0
MySQL-python==1.2.5
psycopg2==2.6.2
pytz==2016.10
requests==2.12.4
six==1.10.0
whitenoise==3.2.2
